Actions now are performed using space and not time
Actions system has ben reworked to use space instead of actions duration. Now every actions will cover a defined number of units (specified in MovementParams SO). While running, jumps will be longer.
Movement is handled by using a rigidbody attached to character control whose velocity is modified at every fixed update. gravity multiplier is set to 0 whenever the character is on the ground. When it must jump gravity multiplier is set to a value and on landing it is set back to 0.