Added transition starting ending a level
These are new transitions that show themselves when starting or ending a level. It's showed a panel witch change the opacity: from black to transparent and from transparent to black.
At the very first second beginning of the transition you can't use the platform bar, but this shouldn't be a problem because we will add a camera preview at the beginning of the level.
At the end of the level there is the finish level transition where you can see Luke continuing moving durning this transition, and this will be to fix by creating a Luke's animation when he die o he reach the arrive.
I've created a prefab to add easily these transitions in a level! When I'll do the rebase, in case I'll have to discard my change at levels and reattach this prefab to each one.